Final Fix

+++ app/src/responses/dish_resp.py
@@ -1,7 +1,10 @@
 import json
 
 import aioredis
+from database import get_db
+from fastapi import Depends
 from fastapi.responses import JSONResponse
+from sqlalchemy.ext.asyncio import AsyncSession
 from src.DAO.dish_DAO import DishDao
 
 r = aioredis.from_url(
@@ -10,7 +13,7 @@ r = aioredis.from_url(
 
 
 class DishResp:
-    async def get_dishes(menu_id, submenu_id):
+    async def get_dishes(menu_id, submenu_id, dish_service):
         dishes = await r.get(
             json.dumps(
                 [menu_id, submenu_id, "dishes"],
@@ -19,7 +22,8 @@ class DishResp:
         if dishes is not None:
             return json.loads(dishes)
         else:
-            dishes = await DishDao.get_dishes(submenu_id)  # type: ignore
+            # type: ignore
+            dishes = await DishDao.get_dishes(submenu_id, dish_service)
             await r.set(
                 json.dumps(
                     [menu_id, submenu_id, "dishes"],
@@ -28,7 +32,7 @@ class DishResp:
             )
             return dishes
 
-    async def get_dish(menu_id, submenu_id, dish_id):
+    async def get_dish(menu_id, submenu_id, dish_id, dish_service):
         dish = await r.get(
             json.dumps(
                 [menu_id, submenu_id, dish_id],
@@ -37,7 +41,7 @@ class DishResp:
         if dish is not None:
             return json.loads(dish)
         else:
-            dish = await DishDao.get_dish(submenu_id, dish_id)
+            dish = await DishDao.get_dish(submenu_id, dish_id, dish_service)
             await r.set(
                 json.dumps(
                     [menu_id, submenu_id, dish_id],
@@ -54,7 +58,7 @@ class DishResp:
             else:
                 return dish
 
-    async def create_dish(menu_id, submenu_id, data):
+    async def create_dish(menu_id, submenu_id, data, dish_service):
         await r.delete(
             json.dumps([menu_id, submenu_id, "dishes"]),
             json.dumps([menu_id, "submenus"]),
@@ -62,12 +66,12 @@ class DishResp:
             menu_id,
             "menus",
         )
-        content = await DishDao.create_dish(submenu_id, data)
+        content = await DishDao.create_dish(submenu_id, data, dish_service)
         return JSONResponse(status_code=201, content=content)
 
-    async def edit_dish(menu_id, submenu_id, dish_id, data):
+    async def edit_dish(menu_id, submenu_id, dish_id, data, dish_service):
         await r.flushall()
-        dish = await DishDao.edit_dish(submenu_id, dish_id, data)
+        dish = await DishDao.edit_dish(submenu_id, dish_id, data, dish_service)
         if dish is None:
             return JSONResponse(
                 status_code=404,
@@ -78,7 +82,19 @@ class DishResp:
         else:
             return dish
 
-    async def delete_dish(menu_id, submenu_id, dish_id):
-        await r.flushall()
-        await DishDao.delete_dish(submenu_id, dish_id)
-        return {"status": True, "message": "The dish has been deleted"}
+    async def delete_dish(menu_id, submenu_id, dish_id, dish_service):
+        dish = await DishDao.delete_dish(submenu_id, dish_id, dish_service)
+        if dish is None:
+            return JSONResponse(
+                status_code=404,
+                content={
+                    "detail": "dish not found",
+                },
+            )
+        else:
+            await r.flushall()
+            return {"status": True, "message": "The dish has been deleted"}
+
+
+async def get_dish_service(db: AsyncSession = Depends(get_db)):
+    return db
